Remove obsolete float64 casting

+++ layout/force_gravity.go
@@ -22,16 +22,16 @@ func NewGravityForce(coeff float64, rule ForceRule) Force {
 
 // Apply calculates the gravity force between two nodes. Satisfies Force interface.
 func (g *GravityForce) Apply(from, to Point) *ForceVector {
-	xx := float64(to.X() - from.X())
-	yy := float64(to.Y() - from.Y())
-	zz := float64(to.Z() - from.Z())
+	xx := to.X() - from.X()
+	yy := to.Y() - from.Y()
+	zz := to.Z() - from.Z()
 
-	r := float64(math.Sqrt(float64(xx*xx) + float64(yy*yy) + float64(zz*zz)))
+	r := math.Sqrt(xx*xx + yy*yy + zz*zz)
 	if r == 0 {
 		r = 10
 	}
 
-	v := g.Coeff * float64(from.Mass()*to.Mass()) / float64(r*r*r)
+	v := g.Coeff * from.Mass() * to.Mass() / (r * r * r)
 	return &ForceVector{
 		DX: (xx * v),
 		DY: (yy * v),
